+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ROM support for starter randomization. (AP world side is still missing)
 * Hold A to advance option.
 * ROM support for randomizing the species that Rowan shows during the intro. (AP world side is still missing)
+### Changed
+* In-game game options option is now validated before generation.
 ### Fixed
 * Non-determinism of generation.
 * Elite Four rematch is triggered only if the game has been completed and stark mountain has been cleared. (untested)